requirements.txt
dj-database-url
cloudinary
django-cloudinary-storage
psycopg2-binary
gunicorn
